+++ subchain/weak/config.yaml
