+++ subchain/weak/honest_miner.py
